frontend/app/server/requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
asyncpg==0.28.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.12.0
psycopg2-binary==2.9.7
